pyproject.toml
[project]
name = "stfdowntown-metagpt"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10.0"
dependencies = [
    "agent-network-test",
    "aiofiles==23.2.1",
    "aiohttp==3.8.6",
    "anthropic==0.47.2",
    "anytree>=2.12.1",
    "beautifulsoup4==4.12.3",
    "boto3~=1.34.69",
    "channels==4.0.0",
    "curl-cffi~=0.7.0",
    "dashscope~=1.19.3",
    "faiss-cpu==1.7.4",
    "fire==0.4.0",
    "fsspec>=2025.5.1",
    "gitignore-parser==0.1.9",
    "gitpython==3.1.40",
    "google-generativeai==0.4.1",
    "grep-ast~=0.3.3",
    "grpcio~=1.67.0",
    "grpcio-status~=1.62.3",
    "grpcio-tools~=1.62.3",
    "gymnasium==0.29.1",
    "htmlmin>=0.1.12",
    "httplib2~=0.22.0",
    "httpx==0.27.2",
    "imap-tools==1.5.0",
    "ipykernel==6.27.1",
    "ipython==8.17.2",
    "ipywidgets==8.1.1",
    "jieba==0.42.1",
    "lancedb==0.4.0",
    "libcst==1.0.1",
    "loguru==0.6.0",
    "meilisearch==0.21.0",
    "nbclient==0.9.0",
    "nbformat==5.9.2",
    "networkx~=3.2.1",
    "numpy~=1.26.4",
    "openai==1.44.1",
    "openpyxl~=3.1.5",
    "pandas==2.1.1",
    "pillow>=11.2.1",
    "playwright>=1.26",
    "pydantic>=2.5.3",
    "pygithub~=2.3",
    "pylint~=3.0.3",
    "python-docx==0.8.11",
    "pyyaml==6.0.2",
    "qdrant-client==1.7.0",
    "qianfan~=0.4.4",
    "rank-bm25==0.2.2",
    "redis~=5.0.0",
    "rich==13.6.0",
    "scikit-learn==1.3.2",
    "semantic-kernel==0.4.3.dev0",
    "setuptools>=65.6.3",
    "socksio~=1.0.0",
    "spark-ai-python~=0.3.30",
    "ta==0.10.2",
    "tenacity==8.2.3",
    "tiktoken==0.9.0",
    "tqdm==4.66.5",
    "tree-sitter~=0.23.2",
    "tree-sitter-python~=0.23.2",
    "typer==0.9.0",
    "typing-extensions==4.12.2",
    "typing-inspect==0.8.0",
    "unidiff==0.7.5",
    "volcengine-python-sdk[ark]~=1.0.94",
    "websocket-client~=1.8.0",
    "websockets>=10.0,<12.0",
    "wrapt==1.15.0",
    "zhipuai~=2.1.5",
]

[tool.uv.sources]
agent-network-test = { git = "https://github.com/WhuAgent/agent-network.git", rev = "ccbd1a961b22bf2d155eb0d583fa43b35d255dc5" }
